+++ main/ad2_settings.h
@@ -23,9 +23,6 @@
 // @brief Define for task stack reporting for tuning.
 //#define AD2_STACK_REPORT
 
-// @brief Firmware version string.
-#define FIRMWARE_VERSION      "AD2IOT-1104"
-
 #if defined(CONFIG_STDK_IOT_CORE)
 #define FIRMWARE_BUILDFLAGS "stsdk"
 // Default netmode N disabled